make AtIndexer __call__ method specific to TreeClass only
- + move indexing code
- add lic header to test file
