+++ pytreeclass/__init__.py
@@ -1,5 +1,20 @@
+# Copyright 2023 PyTreeClass authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from pytreeclass._src.code_build import field, fields
-from pytreeclass._src.tree_base import AtIndexer, TreeClass
+from pytreeclass._src.tree_base import TreeClass
+from pytreeclass._src.tree_index import AtIndexer, BaseKey
 from pytreeclass._src.tree_mask import (
     freeze,
     is_frozen,
@@ -19,7 +34,6 @@ from pytreeclass._src.tree_pprint import (
     tree_summary,
 )
 from pytreeclass._src.tree_util import (
-    BaseKey,
     Partial,
     bcmap,
     is_tree_equal,
@@ -42,22 +56,23 @@ __all__ = (
     "tree_indent",
     "tree_summary",
     "pp_dispatcher",
-    # freeze/unfreeze utils
+    # masking utils
     "is_nondiff",
     "is_frozen",
     "freeze",
     "unfreeze",
     "tree_mask",
     "tree_unmask",
-    # masking and indexing utils
-    "bcmap",
+    # indexing utils
     "AtIndexer",
+    "BaseKey",
+    # tree utils
+    "bcmap",
     "tree_map_with_trace",
     "tree_leaves_with_trace",
     "tree_flatten_with_trace",
     "tree_repr_with_trace",
     "Partial",
-    "BaseKey",
 )
 
 __version__ = "0.4.0"
