+++ pytreeclass/_src/tree_base.py
@@ -17,14 +17,11 @@
 from __future__ import annotations
 
 import abc
-from collections.abc import Callable
 from contextlib import contextmanager
-from typing import Any, Hashable, NamedTuple, TypeVar
+from typing import Any, Hashable, TypeVar
 
 import jax
-import jax.numpy as jnp
 import jax.tree_util as jtu
-import numpy as np
 from typing_extensions import Unpack, dataclass_transform
 
 from pytreeclass._src.code_build import (
@@ -34,6 +31,7 @@ from pytreeclass._src.code_build import (
     field,
     fields,
 )
+from pytreeclass._src.tree_index import AtIndexer
 from pytreeclass._src.tree_pprint import (
     PPSpec,
     attr_value_pp,
@@ -43,12 +41,8 @@ from pytreeclass._src.tree_pprint import (
     tree_str,
 )
 from pytreeclass._src.tree_util import (
-    BaseKey,
-    IsLeafType,
     NamedSequenceKey,
-    NameKey,
     _leafwise_transform,
-    _resolve_where,
     is_tree_equal,
     tree_copy,
     tree_hash,
@@ -58,7 +52,6 @@ T = TypeVar("T", bound=Hashable)
 S = TypeVar("S")
 PyTree = Any
 EllipsisType = type(Ellipsis)
-_no_initializer = object()
 
 
 # allow methods in mutable context to be called without raising `AttributeError`
@@ -107,310 +100,7 @@ def _register_treeclass(klass: type[T]) -> type[T]:
     return klass
 
 
-class AtIndexer(NamedTuple):
-    """Index a pytree at a given path using a path or mask.
-
-    Args:
-        tree: pytree to index
-        where: one of the following:
-            - `str` for mapping keys or class attributes.
-            - `int` for positional indexing for sequences.
-            - `...` to select all leaves.
-            - a boolean mask of the same structure as the tree
-            - `re.Pattern` to index all keys matching a regex pattern.
-            - an instance of `BaseKey` with custom logic to index a pytree.
-            - a tuple of the above to index multiple keys at the same level.
-
-    Example:
-        >>> # use `AtIndexer` on a pytree (e.g. dict,list,tuple,etc.)
-        >>> import pytreeclass as pytc
-        >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
-        >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0"].get()
-        {'level1_0': {'level2_0': 100, 'level2_1': None}, 'level1_1': None}
-        >>> # get multiple keys at once at the same level
-        >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0", "level2_1"].get()
-        {'level1_0': {'level2_0': 100, 'level2_1': 200}, 'level1_1': None}
-        >>> # get with a mask
-        >>> mask = {"level1_0": {"level2_0": True, "level2_1": False}, "level1_1": True}
-        >>> pytc.AtIndexer(tree).at[mask].get()
-        {'level1_0': {'level2_0': 100, 'level2_1': None}, 'level1_1': 300}
-
-    Example:
-        >>> # use `AtIndexer` in a class
-        >>> import jax.tree_util as jtu
-        >>> import pytreeclass as pytc
-        >>> @jax.tree_util.register_pytree_with_keys_class
-        ... class Tree:
-        ...    def __init__(self, a, b):
-        ...        self.a = a
-        ...        self.b = b
-        ...    def tree_flatten_with_keys(self):
-        ...        kva = (jtu.GetAttrKey("a"), self.a)
-        ...        kvb = (jtu.GetAttrKey("b"), self.b)
-        ...        return (kva, kvb), None
-        ...    @classmethod
-        ...    def tree_unflatten(cls, aux_data, children):
-        ...        return cls(*children)
-        ...    @property
-        ...    def at(self):
-        ...        return pytc.AtIndexer(self)
-        ...    def __repr__(self) -> str:
-        ...        return f"{self.__class__.__name__}(a={self.a}, b={self.b})"
-        >>> Tree(1, 2).at["a"].get()
-        Tree(a=1, b=None)
-    """
-
-    tree: PyTree
-    where: tuple[BaseKey | PyTree] | tuple[()] = ()
-
-    def __getitem__(self, where: Any) -> AtIndexer:
-        return AtIndexer(self.tree, (*self.where, where))
-
-    def __getattr__(self, name: str) -> AtIndexer:
-        """Support nested indexing"""
-        if name == "at":
-            # pass the current tree and the current path to the next `.at`
-            return AtIndexer(tree=self.tree, where=self.where)
-
-        raise AttributeError(f"`{self!r}` has no attribute {name!r}.")
-
-    def get(self, *, is_leaf: IsLeafType = None) -> PyTree:
-        """Get the leaf values at the specified location.
-
-        Args:
-            is_leaf: a predicate function to determine if a value is a leaf.
-
-        Returns:
-            A _new_ pytree of leaf values at the specified location, with the
-            non-selected leaf values set to None if the leaf is not an array.
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
-            >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0"].get()
-            {'level1_0': {'level2_0': 100, 'level2_1': None}, 'level1_1': None}
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> class Tree(pytc.TreeClass):
-            ...     a: int
-            ...     b: int
-            >>> tree = Tree(a=1, b=2)
-            >>> # get `a` and return a new instance
-            >>> # with `None` for all other leaves
-            >>> tree.at['a'].get()
-            Tree(a=1, b=None)
-        """
-        where = _resolve_where(self.tree, self.where, is_leaf)
-
-        def leaf_get(leaf: Any, where: Any):
-            if isinstance(where, (jax.Array, np.ndarray)) and where.ndim != 0:
-                return leaf[jnp.where(where)]
-            return leaf if where else None
-
-        return jtu.tree_map(leaf_get, self.tree, where, is_leaf=is_leaf)
-
-    def set(self, set_value: Any, *, is_leaf: IsLeafType = None):
-        """Set the leaf values at the specified location.
-
-        Args:
-            set_value: the value to set at the specified location.
-            is_leaf: a predicate function to determine if a value is a leaf.
-
-        Returns:
-            A pytree with the leaf values at the specified location
-            set to `set_value`.
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
-            >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0"].set('SET')
-            {'level1_0': {'level2_0': 'SET', 'level2_1': 200}, 'level1_1': 300}
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> class Tree(pytc.TreeClass):
-            ...     a: int
-            ...     b: int
-            >>> tree = Tree(a=1, b=2)
-            >>> # set `a` and return a new instance
-            >>> # with all other leaves unchanged
-            >>> tree.at['a'].set(100)
-            Tree(a=100, b=2)
-        """
-        where = _resolve_where(self.tree, self.where, is_leaf)
-
-        def leaf_set(leaf: Any, where: Any, set_value: Any):
-            if isinstance(where, (jax.Array, np.ndarray)):
-                return jnp.where(where, set_value, leaf)
-            return set_value if where else leaf
-
-        if jtu.tree_structure(self.tree, is_leaf) == jtu.tree_structure(set_value):
-            # do not broadcast set_value if it is a pytree of same structure
-            # for example tree.at[where].set(tree2) will set all tree leaves
-            # to tree2 leaves if tree2 is a pytree of same structure as tree
-            # instead of making each leaf of tree a copy of tree2
-            # is design is similar to `numpy` design `Array.at[...].set(Array)`
-            return jtu.tree_map(leaf_set, self.tree, where, set_value, is_leaf=is_leaf)
-
-        # set_value is broadcasted to tree leaves
-        # for example tree.at[where].set(1) will set all tree leaves to 1
-        partial_leaf_set = lambda leaf, where: leaf_set(leaf, where, set_value)
-        return jtu.tree_map(partial_leaf_set, self.tree, where, is_leaf=is_leaf)
-
-    def apply(self, func: Callable[[Any], Any], *, is_leaf: IsLeafType = None):
-        """Apply a function to the leaf values at the specified location.
-
-        Args:
-            func: the function to apply to the leaf values.
-            is_leaf: a predicate function to determine if a value is a leaf.
-
-        Returns:
-            A pytree with the leaf values at the specified location set to
-            the result of applying `func` to the leaf values.
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
-            >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0"].apply(lambda _: 'SET')
-            {'level1_0': {'level2_0': 'SET', 'level2_1': 200}, 'level1_1': 300}
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> class Tree(pytc.TreeClass):
-            ...     a: int
-            ...     b: int
-            >>> tree = Tree(a=1, b=2)
-            >>> # apply to `a` and return a new instance
-            >>> # with all other leaves unchanged
-            >>> tree.at['a'].apply(lambda _: 100)
-            Tree(a=100, b=2)
-        """
-        where = _resolve_where(self.tree, self.where, is_leaf)
-
-        def leaf_apply(leaf: Any, where: bool):
-            if isinstance(where, (jax.Array, np.ndarray)):
-                return jnp.where(where, func(leaf), leaf)
-            return func(leaf) if where else leaf
-
-        return jtu.tree_map(leaf_apply, self.tree, where, is_leaf=is_leaf)
-
-    def scan(
-        self,
-        func: Callable[[Any, S], tuple[Any, S]],
-        state: S,
-        *,
-        is_leaf: IsLeafType = None,
-    ) -> tuple[PyTree, S]:
-        """Apply a function to the leaf values at the specified location defined
-        by the mask while carrying a state.
-
-        Args:
-            func: the function to apply to the leaf values. the function accepts
-                a leaf value and a state and returns a tuple of the new leaf
-                value and updated state.
-            state: the initial state to carry.
-            is_leaf: a predicate function to determine if a value is a leaf. for
-                example, `lambda x: isinstance(x, list)` will treat all lists
-                as leaves and will not recurse into list items.
-
-        Returns:
-            A tuple of a pytree with the leaf values at the specified location
-            set to the result of applying `func` to the leaf values and the
-            new state.
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
-            >>> def scan_func(leaf, state):
-            ...     return 'SET', state + 1
-            >>> init_state = 0
-            >>> pytc.AtIndexer(tree).at["level1_0"].at["level2_0"].scan(scan_func, state=init_state)
-            ({'level1_0': {'level2_0': 'SET', 'level2_1': 200}, 'level1_1': 300}, 1)
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> from typing import NamedTuple
-            >>> class State(NamedTuple):
-            ...     func_evals: int = 0
-            >>> class Tree(pytc.TreeClass):
-            ...     a: int
-            ...     b: int
-            ...     c: int
-            >>> tree = Tree(a=1, b=2, c=3)
-            >>> def scan_func(leaf, state: State):
-            ...     state = State(state.func_evals + 1)
-            ...     return leaf + 1, state
-            >>> # apply to `a` and `b` and return a new instance with all other
-            >>> # leaves unchanged and the new state that counts the number of
-            >>> # function evaluations
-            >>> tree.at['a','b'].scan(scan_func, state=State())
-            (Tree(a=2, b=3, c=3), State(func_evals=2))
-
-        Note:
-            - `scan` applies a binary `func` to the leaf values while carrying
-                a state and returning a final state and tree leaves with the
-                the `func` applied to them. While `reduce` applies a binary
-                `func` to the leaf values while carrying a state and returning
-                a single value.
-        """
-        where = _resolve_where(self.tree, self.where, is_leaf)
-
-        running_state = state
-
-        def stateless_func(leaf):
-            nonlocal running_state
-            leaf, running_state = func(leaf, running_state)
-            return leaf
-
-        def leaf_apply(leaf: Any, where: bool):
-            if isinstance(where, (jax.Array, np.ndarray)):
-                return jnp.where(where, stateless_func(leaf), leaf)
-            return stateless_func(leaf) if where else leaf
-
-        out = jtu.tree_map(leaf_apply, self.tree, where, is_leaf=is_leaf)
-        return out, running_state
-
-    def reduce(
-        self,
-        func: Callable[[Any, Any], Any],
-        *,
-        initializer: Any = _no_initializer,
-        is_leaf: IsLeafType = None,
-    ) -> Any:
-        """Reduce the leaf values at the specified location.
-
-        Args:
-            func: the function to reduce the leaf values.
-            initializer: the initializer value for the reduction.
-            is_leaf: a predicate function to determine if a value is a leaf.
-
-        Returns:
-            The result of reducing the leaf values at the specified location.
-
-        Note:
-            - If `initializer` is not specified, the first leaf value is used as
-                the initializer.
-            - `reduce` applies a binary `func` to each leaf values while accumulating
-                a state a returns the final result. while `scan` applies `func` to each
-                leaf value while carrying a state and returns the final state and
-                the leaves of the tree with the result of applying `func` to each leaf.
-
-        Example:
-            >>> import pytreeclass as pytc
-            >>> class Tree(pytc.TreeClass):
-            ...     a: int
-            ...     b: int
-            >>> tree = Tree(a=1, b=2)
-            >>> tree.at[...].reduce(lambda a, b: a + b, initializer=0)
-            3
-        """
-        where = _resolve_where(self.tree, self.where, is_leaf)
-        tree = self.tree.at[where].get(is_leaf=is_leaf)  # type: ignore
-        if initializer is _no_initializer:
-            return jtu.tree_reduce(func, tree)
-        return jtu.tree_reduce(func, tree, initializer)
-
+class TreeClassIndexer(AtIndexer):
     def __call__(self, *a, **k) -> tuple[Any, PyTree]:
         """
         Call the function at the specified location and return a **copy**
@@ -438,7 +128,7 @@ class AtIndexer(NamedTuple):
             - Use .at["method_name"](*, **) to call a method that mutates the instance.
         """
 
-        def recursive_getattr(tree: Any, where: tuple[NameKey, ...]):
+        def recursive_getattr(tree: Any, where: tuple[str, ...]):
             if not isinstance(where[0], str):
                 raise TypeError(f"Expected string, got {type(where[0])!r}.")
             if len(where) == 1:
@@ -597,7 +287,7 @@ class TreeClass(metaclass=TreeClassMeta):
         getattr(object, "__delattr__")(self, key)
 
     @property
-    def at(self) -> AtIndexer:
+    def at(self) -> TreeClassIndexer:
         """Immutable out-of-place indexing
 
         - `.at[***].get()`:
@@ -648,7 +338,7 @@ class TreeClass(metaclass=TreeClassMeta):
             - `pytree.at[*].at[**]` is equivalent to selecting pytree.*.**
             - `pytree.at[*, **]` is equivalent selecting pytree.* and pytree.**
         """
-        return AtIndexer(self)
+        return TreeClassIndexer(self)
 
     def __repr__(self) -> str:
         return tree_repr(self)
